Split workspace state Playwright flows
