+++ E2E/playwright/tests/memories.spec.ts
@@ -1,0 +1,55 @@
+import { test, expect } from '@playwright/test';
+import { clickSidebarTool, harnessURL } from './harness-helpers';
+
+test('mock harness shows memories from sidebar and command palette', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await expect(page.getByTestId('project-memories-status')).toHaveText('2 memories');
+  await clickSidebarTool(page, 'memories-button');
+
+  await expect(page.getByTestId('memories-pane')).toBeVisible();
+  await expect(page.getByTestId('memories-subtitle')).toHaveText('1 global memory · 1 project memory');
+  await expect(page.getByTestId('memory-item')).toHaveCount(2);
+  await expect(page.getByTestId('memory-title').first()).toHaveText('Preferences');
+  await expect(page.getByTestId('memory-path').first()).toHaveText('memories/preferences.md');
+  await expect(page.getByTestId('memory-delete')).toHaveCount(1);
+  await expect(page.getByTestId('memories-add')).toBeVisible();
+
+  await clickSidebarTool(page, 'command-palette-button');
+  await page.getByLabel('Search commands').fill('>memories');
+  await page.getByTestId('command-palette-result').filter({ hasText: 'Memories' }).click();
+
+  await expect(page.getByTestId('memories-pane')).toHaveCount(0);
+
+  await clickSidebarTool(page, 'command-palette-button');
+  await page.getByLabel('Search commands').fill('>save');
+  await expect(page.getByTestId('command-palette-result')).toHaveCount(1);
+  await expect(page.getByTestId('command-palette-result')).toContainText('Add memory');
+  await page.getByTestId('command-palette-result').click();
+
+  await expect(page.getByLabel('Message')).toHaveValue('/remember ');
+  await page.getByLabel('Message').fill('/remember Prefer small reviewable commits');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByText('Saved memory: Prefer Small Reviewable Commits. It will be included as background context in future turns.')).toBeVisible();
+  await expect(page.getByTestId('project-memories-status')).toHaveText('3 memories');
+  await expect(page.getByTestId('top-bar-title')).toHaveText('Memory: Prefer Small Reviewable Commits');
+
+  await clickSidebarTool(page, 'memories-button');
+  await expect(page.getByTestId('memories-pane')).toBeVisible();
+  await expect(page.getByTestId('memories-subtitle')).toHaveText('2 global memories · 1 project memory');
+  await expect(page.getByTestId('memory-item')).toHaveCount(3);
+  await expect(page.getByTestId('memory-title').first()).toHaveText('Prefer Small Reviewable Commits');
+  await expect(page.getByTestId('memory-path').first()).toContainText('memories/manual-');
+  await expect(page.getByTestId('memory-delete')).toHaveCount(2);
+
+  await page.getByTestId('memory-delete').first().click();
+
+  await expect(page.getByText('Forgot memory: Prefer Small Reviewable Commits. It will no longer be included as background context.')).toBeVisible();
+  await expect(page.getByTestId('project-memories-status')).toHaveText('2 memories');
+  await expect(page.getByTestId('top-bar-title')).toHaveText('Forgot memory: Prefer Small Reviewable Commits');
+  await expect(page.getByTestId('memories-subtitle')).toHaveText('1 global memory · 1 project memory');
+  await expect(page.getByTestId('memory-item')).toHaveCount(2);
+  await expect(page.getByTestId('memory-title').first()).toHaveText('Preferences');
+  await expect(page.getByTestId('memory-delete')).toHaveCount(1);
+});
